python-postgresql/requirements.txt
pyyaml==6.0
grpcio==1.50.0
